fixup! Redirect URLs with a Trailing Slash (code 301)
